fix(favicons): cache shows wrong Google Docs icons
When this plugin fetches favicon URLs, it uses a basic Map to cache
the favicon response bodies in memory. Previously this Map was keyed
using the hostname of the URL in the block content. However, this caused
an incompatibility with hard-coded path-specific exceptions that this
plugin implements for Google Docs.

So, if you have the following Logseq page:

```md
- https://docs.google.com/document/xxx
- https://docs.google.com/spreadsheets/xxx
- https://docs.google.com/presentation/xxx
```

The cache would side-step the path-specific checks, and all the links
would display the same icon.

This change shifts the cache around just the fetch of the favicon, and
uses the favicon's URL as the cache key. Note this means the cache is no
longer covers the conditionals in getFaviconData, but this should be a
minimal tradeoff for displaying the correct icon.